fix(api): harden rate limiter against header spoofing + clippy cleanup
Next-level review pass (cargo clippy --fix + security audit of auth/middleware).

SECURITY (real bug):
- middleware/rate_limit.rs: the IP fallback key was derived from the
  client-controlled 'x-forwarded-for' header (and 'user-agent'), so an
  attacker could rotate either per request to fully bypass the limiter.
  The server binds directly with no trusted-proxy config, so the forwarded
  header is untrusted. Now derive the key from the real peer socket
  address via axum ConnectInfo<SocketAddr> (wired through
  into_make_service_with_connect_info in lib.rs). Authenticated tenants
  still key by tenant id; unauthenticated traffic keys by peer IP.

CLIPPY (mechanical cleanups, auto-fixed + manual):
- cargo clippy --fix: redundant closures, collapsible ifs, map().flatten()
  on Option, identity maps (spark_repo_impl.rs x2), in idempotency.rs,
  ip_allowlist.rs, spark.rs, spark_service.rs, vault_service.rs,
  shopify_service.rs, tempo_refresh_worker.rs.
- clippy.toml: add too-many-arguments-threshold = 12, documenting that
  DI constructors (PivotService::new, log_audit) intentionally take many
  collaborators.

Verified: cargo clippy --workspace --all-targets = 0 warnings/errors;
cargo test --workspace = exit 0 (all crates, 0 failed); pnpm run test
187/187 passing (TS layer unaffected).

+++ crates/ataqu-application/src/tempo_refresh_worker.rs
@@ -56,7 +56,7 @@ pub async fn refresh_expiring_tokens(
 
         match refresh_token_for_provider(&client, provider, &refresh_token).await {
             Ok((new_access_token, new_expires_in)) => {
-                let new_expiry = Utc::now() + Duration::seconds(new_expires_in as i64);
+                let new_expiry = Utc::now() + Duration::seconds(new_expires_in);
                 let update_stmt = Statement::from_sql_and_values(
                     DbBackend::Postgres,
                     "UPDATE collab_ops.bookings SET oauth_access_token = $1, oauth_token_expires_at = $2 WHERE id = $3",
